Authentication/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
coverage==7.0.1
cryptography==39.0.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.3
django-allauth==0.52.0
django-bootstrap-form==3.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-heroku==0.3.1
django-npm==1.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.2
git-lfs==1.6
h3==3.7.6
httpie==3.2.1
idna==3.4
loader==2017.9.11
multidict==6.0.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.24.1
oauthlib==3.2.2
optional-django==0.3.0
pbr==5.11.1
Pillow==9.3.0
pipenv==2022.12.19
platformdirs==2.6.0
psycopg2==2.9.5
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PySocks==1.7.1
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.6
react==4.3.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.22.0
rich==13.1.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stevedore==4.1.1
timezonefinder==6.1.9
toml==0.10.2
types-toml==0.10.8.1
tzdata==2022.6
unicode==2.9
Unidecode==1.3.6
urllib3==1.26.13
vboxapi==1.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webpack==6.0.0
whitenoise==6.3.0
